+++ src/libraries/JBRulesetMetadataResolver.sol
@@ -5,18 +5,18 @@ import {JBRuleset} from "./../structs/JBRuleset.sol";
 import {JBRulesetMetadata} from "./../structs/JBRulesetMetadata.sol";
 
 library JBRulesetMetadataResolver {
-    function reservedRate(JBRuleset memory ruleset) internal pure returns (uint256) {
-        return uint256(uint16(ruleset.metadata >> 4));
+    function reservedRate(JBRuleset memory ruleset) internal pure returns (uint16) {
+        return uint16(ruleset.metadata >> 4);
     }
 
-    function redemptionRate(JBRuleset memory ruleset) internal pure returns (uint256) {
+    function redemptionRate(JBRuleset memory ruleset) internal pure returns (uint16) {
         // Redemption rate is a number 0-10000.
-        return uint256(uint16(ruleset.metadata >> 20));
+        return uint16(ruleset.metadata >> 20);
     }
 
-    function baseCurrency(JBRuleset memory ruleset) internal pure returns (uint256) {
+    function baseCurrency(JBRuleset memory ruleset) internal pure returns (uint32) {
         // Currency is a number 0-4294967296.
-        return uint256(uint32(ruleset.metadata >> 36));
+        return uint32(ruleset.metadata >> 36);
     }
 
     function pausePay(JBRuleset memory ruleset) internal pure returns (bool) {
@@ -71,8 +71,8 @@ library JBRulesetMetadataResolver {
         return address(uint160(ruleset.metadata >> 80));
     }
 
-    function metadata(JBRuleset memory ruleset) internal pure returns (uint256) {
-        return uint256(uint16(ruleset.metadata >> 240));
+    function metadata(JBRuleset memory ruleset) internal pure returns (uint8) {
+        return uint8(ruleset.metadata >> 240);
     }
 
     /// @notice Pack the funding cycle metadata.
@@ -82,13 +82,13 @@ library JBRulesetMetadataResolver {
         // version 1 in the bits 0-3 (4 bits).
         packed = 1;
         // reserved rate in bits 4-19 (16 bits).
-        packed |= rulesetMetadata.reservedRate << 4;
+        packed |= uint256(rulesetMetadata.reservedRate) << 4;
         // redemption rate in bits 20-35 (16 bits).
         // redemption rate is a number 0-10000.
-        packed |= rulesetMetadata.redemptionRate << 20;
+        packed |= uint256(rulesetMetadata.redemptionRate) << 20;
         // base currency in bits 36-67 (32 bits).
         // base currency is a number 0-16777215.
-        packed |= rulesetMetadata.baseCurrency << 36;
+        packed |= uint256(rulesetMetadata.baseCurrency) << 36;
         // pause pay in bit 68.
         if (rulesetMetadata.pausePay) packed |= 1 << 68;
         // pause credit transfers in bit 69.
@@ -115,8 +115,8 @@ library JBRulesetMetadataResolver {
         if (rulesetMetadata.useDataHookForRedeem) packed |= 1 << 79;
         // data source address in bits 80-239.
         packed |= uint256(uint160(address(rulesetMetadata.dataHook))) << 80;
-        // metadata in bits 240-255 (16 bits).
-        packed |= rulesetMetadata.metadata << 240;
+        // metadata in bits 240-247 (8 bits).
+        packed |= uint256(rulesetMetadata.metadata) << 240;
     }
 
     /// @notice Expand the funding cycle metadata.
