+++ src/structs/JBCurrencyAmount.sol
@@ -5,6 +5,6 @@ pragma solidity ^0.8.0;
 /// @custom:member currency The currency. By convention, this is `uint32(uint160(tokenAddress))` for tokens, or a
 /// constant ID from e.g. `JBCurrencyIds` for other currencies.
 struct JBCurrencyAmount {
-    uint256 amount;
-    uint256 currency;
+    uint224 amount;
+    uint32 currency;
 }
