+++ test/units/static/JBRulesetMetadataResolver/TestSetRedemptionRateTo.sol
@@ -9,8 +9,8 @@ contract TestSetRedemptionRateTo_Local is JBTest {
     function setUp() external {}
 
     function testFuzzEnsureCorrectlyPackedBits(
-        uint256 _fuzzReservedRate,
-        uint256 _fuzzRedemptionRate,
+        uint16 _fuzzReservedRate,
+        uint16 _fuzzRedemptionRate,
         uint8 _fuzzMetadata
     )
         external
@@ -19,8 +19,8 @@ contract TestSetRedemptionRateTo_Local is JBTest {
 
         address _hookAddress = makeAddr("someting");
 
-        _fuzzReservedRate = bound(_fuzzReservedRate, 0, JBConstants.MAX_RESERVED_RATE);
-        _fuzzRedemptionRate = bound(_fuzzRedemptionRate, 0, JBConstants.MAX_REDEMPTION_RATE);
+        _fuzzReservedRate = uint16(bound(_fuzzReservedRate, 0, JBConstants.MAX_RESERVED_RATE));
+        _fuzzRedemptionRate = uint16(bound(_fuzzRedemptionRate, 0, JBConstants.MAX_REDEMPTION_RATE));
 
         JBRulesetMetadata memory _rulesMetadata = JBRulesetMetadata({
             reservedRate: _fuzzReservedRate,
